move accessor test to integration test
