+++ src/API/WesternStatesWater.WestDaat.Tests.AccessorTests/ApplicationAccessorTests.cs
@@ -1,5 +1,4 @@
 using Microsoft.EntityFrameworkCore;
-using System.Reflection;
 using WesternStatesWater.WestDaat.Accessors;
 using WesternStatesWater.WestDaat.Common.DataContracts;
 using WesternStatesWater.WestDaat.Database.EntityFramework;
@@ -27,82 +26,6 @@ public class ApplicationAccessorTests : AccessorTestBase
         );
 
         _westDaatDb = westDaatDbContextFactory.Create();
-    }
-
-    [TestMethod]
-
-    public async Task Load_GetFullApplicationDetails_SucceedsForEveryRequestType()
-    {
-        // Arrange
-        var user = new UserFaker().Generate();
-        var org = new OrganizationFaker().Generate();
-
-        var application = new WaterConservationApplicationFaker(user, org).Generate();
-        await _westDaatDb.WaterConservationApplications.AddAsync(application);
-
-        var estimate = new WaterConservationApplicationEstimateFaker(application).Generate();
-        await _westDaatDb.WaterConservationApplicationEstimates.AddAsync(estimate);
-
-        var locations = new WaterConservationApplicationEstimateLocationFaker(estimate).Generate(3);
-        await _westDaatDb.WaterConservationApplicationEstimateLocations.AddRangeAsync(locations);
-
-        foreach (var location in locations)
-        {
-            var consumptiveUses = new WaterConservationApplicationEstimateLocationConsumptiveUseFaker(location).Generate(2);
-            await _westDaatDb.WaterConservationApplicationEstimateLocationConsumptiveUses.AddRangeAsync(consumptiveUses);
-        }
-
-        var submission = new WaterConservationApplicationSubmissionFaker(application).Generate();
-        await _westDaatDb.WaterConservationApplicationSubmissions.AddAsync(submission);
-
-        var documents = new WaterConservationApplicationDocumentsFaker(application).Generate(2);
-        await _westDaatDb.WaterConservationApplicationDocuments.AddRangeAsync(documents);
-
-        await _westDaatDb.SaveChangesAsync();
-
-        // Act
-        var applicantRequest = new ApplicantConservationApplicationLoadRequest
-        {
-            ApplicationId = application.Id
-        };
-        var reviewerRequest = new ReviewerConservationApplicationLoadRequest
-        {
-            ApplicationId = application.Id
-        };
-
-        var applicantResponse = (ApplicantConservationApplicationLoadResponse)await _accessor.Load(applicantRequest);
-        var reviewerResponse = (ReviewerConservationApplicationLoadResponse)await _accessor.Load(reviewerRequest);
-
-        // Assert
-
-        // ExcludingMissingMembers is used to ignore properties that are intentionally not included in the response
-        // i.e. virtual navigation properties and parent id properties
-        applicantResponse.Should().NotBeNull();
-        applicantResponse.Application.Should().BeEquivalentTo(application, options => options.ExcludingMissingMembers());
-        // the following assertions are redundant given that `BeEquivalentTo` checks properties recursively,
-        // but are included for completeness just in case something changes in the future
-        applicantResponse.Application.Estimate.Should().BeEquivalentTo(estimate, options => options.ExcludingMissingMembers());
-        applicantResponse.Application.Estimate.Locations.Should().BeEquivalentTo(locations, options => options.ExcludingMissingMembers());
-        applicantResponse.Application.Estimate.Locations.SelectMany(l => l.ConsumptiveUses).Should().BeEquivalentTo(locations.SelectMany(l => l.ConsumptiveUses), options => options.ExcludingMissingMembers());
-        applicantResponse.Application.Submission.Should().BeEquivalentTo(submission, options => options.ExcludingMissingMembers());
-        applicantResponse.Application.SupportingDocuments.Should().BeEquivalentTo(documents, options => options.ExcludingMissingMembers());
-
-        reviewerResponse.Should().NotBeNull();
-        reviewerResponse.Application.Should().BeEquivalentTo(application, options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.Estimate.Should().BeEquivalentTo(estimate, options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.Estimate.Locations.Should().BeEquivalentTo(locations, options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.Estimate.Locations.SelectMany(l => l.ConsumptiveUses).Should().BeEquivalentTo(locations.SelectMany(l => l.ConsumptiveUses), options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.Submission.Should().BeEquivalentTo(submission, options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.SupportingDocuments.Should().BeEquivalentTo(documents, options => options.ExcludingMissingMembers());
-        reviewerResponse.Application.Notes.Should().BeNull(); // not implemented yet
-
-        // verify that all possible request types are accounted for
-        var requests = Assembly.GetAssembly(typeof(ApplicationLoadSingleRequestBase))?
-            .GetTypes()
-            .Where(t => t.IsClass && !t.IsAbstract && t.IsSubclassOf(typeof(ApplicationLoadSingleRequestBase)))
-            .ToArray();
-
-        requests.Length.Should().Be(2, because: "this test currently only accounts for the specified number of request types. If more request types are added, this test should then be updated.");
     }
 
     [TestMethod]
